Update test runner
- Improve the simulator testing and real device testing stability.

- Fix the wrong detection for latest supported iOS version when installs multiple Xcode on the host.

- Support passing keychain_path in signing_options

+++ xctestrunner/test_runner/xctest_session.py
@@ -154,8 +154,12 @@ class XctestSession(object):
             xcode_info_util.GetXcodeVersionNumber())
       elif test_type == ios_constants.TestType.XCTEST:
         self._dummy_project_obj = dummy_project.DummyProject(
-            app_under_test_dir, test_bundle_dir, self._sdk,
-            ios_constants.TestType.XCTEST, self._work_dir)
+            app_under_test_dir,
+            test_bundle_dir,
+            self._sdk,
+            ios_constants.TestType.XCTEST,
+            self._work_dir,
+            keychain_path=signing_options.get('keychain_path') or None)
         self._dummy_project_obj.GenerateDummyProject()
       elif test_type == ios_constants.TestType.LOGIC_TEST:
         self._logic_test_bundle = test_bundle_dir
@@ -299,16 +303,20 @@ def _PrepareBundles(working_dir, app_under_test_path, test_bundle_path):
       raise ios_errors.IllegalArgumentError(
           'The app under test %s should be with .app or .ipa extension.'
           % app_under_test_path)
-    if app_under_test_path.endswith('.ipa'):
-      app_under_test_dir = bundle_util.ExtractApp(
-          app_under_test_path, working_dir)
-    elif not os.path.abspath(app_under_test_path).startswith(working_dir):
-      # Only copies the app under test if it is not in working directory.
-      app_under_test_dir = os.path.join(
-          working_dir, os.path.basename(app_under_test_path))
-      shutil.copytree(app_under_test_path, app_under_test_dir)
-    else:
-      app_under_test_dir = app_under_test_path
+
+    app_under_test_dir = os.path.join(
+        working_dir,
+        os.path.splitext(os.path.basename(app_under_test_path))[0] + '.app')
+    if not os.path.exists(app_under_test_dir):
+      if app_under_test_path.endswith('.ipa'):
+        extract_app_under_test_dir = bundle_util.ExtractApp(
+            app_under_test_path, working_dir)
+        shutil.move(extract_app_under_test_dir, app_under_test_dir)
+      elif not os.path.abspath(app_under_test_path).startswith(working_dir):
+        # Only copies the app under test if it is not in working directory.
+        shutil.copytree(app_under_test_path, app_under_test_dir)
+      else:
+        app_under_test_dir = app_under_test_path
 
   if not os.path.exists(test_bundle_path):
     raise ios_errors.IllegalArgumentError(
@@ -319,16 +327,20 @@ def _PrepareBundles(working_dir, app_under_test_path, test_bundle_path):
     raise ios_errors.IllegalArgumentError(
         'The test bundle %s should be with .xctest, .ipa or .zip extension.'
         % test_bundle_path)
-  if test_bundle_path.endswith('.ipa') or test_bundle_path.endswith('.zip'):
-    test_bundle_dir = bundle_util.ExtractTestBundle(
-        test_bundle_path, working_dir)
-  elif not os.path.abspath(test_bundle_path).startswith(working_dir):
-    # Only copies the test bundle if it is not in working directory.
-    test_bundle_dir = os.path.join(working_dir,
-                                   os.path.basename(test_bundle_path))
-    shutil.copytree(test_bundle_path, test_bundle_dir)
-  else:
-    test_bundle_dir = test_bundle_path
+
+  test_bundle_dir = os.path.join(
+      working_dir,
+      os.path.splitext(os.path.basename(test_bundle_path))[0] + '.xctest')
+  if not os.path.exists(test_bundle_dir):
+    if test_bundle_path.endswith('.ipa') or test_bundle_path.endswith('.zip'):
+      extract_test_bundle_dir = bundle_util.ExtractTestBundle(
+          test_bundle_path, working_dir)
+      shutil.move(extract_test_bundle_dir, test_bundle_dir)
+    elif not os.path.abspath(test_bundle_path).startswith(working_dir):
+      # Only copies the test bundle if it is not in working directory.
+      shutil.copytree(test_bundle_path, test_bundle_dir)
+    else:
+      test_bundle_dir = test_bundle_path
 
   return app_under_test_dir, test_bundle_dir
 
@@ -388,7 +400,7 @@ def _FinalizeTestType(
 def _DetectTestType(test_bundle_dir):
   """Detects if the test bundle is XCUITest or XCTest."""
   test_bundle_exec_path = os.path.join(
-      test_bundle_dir, os.path.basename(test_bundle_dir).split('.')[0])
+      test_bundle_dir, os.path.splitext(os.path.basename(test_bundle_dir))[0])
   output = subprocess.check_output(['nm', test_bundle_exec_path])
   if 'XCUIApplication' in output:
     return ios_constants.TestType.XCUITEST
